refactor:  로또 유효성 검사 파일로 분할
